Conflicts:
	abonapp/forms.py
	abonapp/migrations/0002_auto_20180808_1448.py
	abonapp/models.py
	abonapp/templates/abonapp/group_list.html
	abonapp/views.py
	clientsideapp/views.py
	djing/local_settings.py.example
	djing/settings.py
	djing/urls.py
	locale/ru/LC_MESSAGES/django.po
	periodic.py

Changes to be committed:
	modified:   abonapp/forms.py
	modified:   abonapp/locale/ru/LC_MESSAGES/django.po
	new file:   abonapp/migrations/0001_squashed_0008_auto_20181115_1206.py
	modified:   abonapp/migrations/0002_auto_20180808_1448.py
	modified:   abonapp/models.py
	new file:   abonapp/tasks.py
	modified:   abonapp/templates/abonapp/debtors.html
	modified:   abonapp/templates/abonapp/editAbon.html
	modified:   abonapp/templates/abonapp/group_list.html
	modified:   abonapp/templates/abonapp/invoiceForPayment.html
	modified:   abonapp/templates/abonapp/modal_phonebook.html
	modified:   abonapp/templates/abonapp/payHistory.html
	modified:   abonapp/templates/abonapp/peoples.html
	modified:   abonapp/templates/abonapp/service.html
	modified:   abonapp/views.py
	modified:   accounts_app/forms.py
	modified:   accounts_app/locale/ru/LC_MESSAGES/django.po
	new file:   accounts_app/migrations/0004_userprofile_flags.py
	modified:   accounts_app/models.py
	modified:   accounts_app/templates/accounts/acc_list.html
	modified:   accounts_app/templates/accounts/ext.htm
	modified:   accounts_app/templates/accounts/index.html
	modified:   accounts_app/templates/accounts/settings/ext.htm
	renamed:    accounts_app/templates/accounts/userprofile_form.html -> accounts_app/templates/accounts/settings/userprofile_form.html
	modified:   accounts_app/urls.py
	modified:   accounts_app/views.py
	new file:   agent/netflow/djing_flow.tar.gz
	modified:   agent/netflow/netflow_handler.py
	modified:   agent/netflow/start_netflow.sh
	deleted:    chatbot/admin.py
	deleted:    chatbot/apps.py
	deleted:    chatbot/locale/ru/LC_MESSAGES/django.po
	deleted:    chatbot/migrations/0001_initial.py
	deleted:    chatbot/migrations/0002_auto_20180808_1236.py
	deleted:    chatbot/models.py
	deleted:    chatbot/telebot.py
	deleted:    chatbot/views.py
	modified:   clientsideapp/views.py
	modified:   devapp/base_intr.py
	modified:   devapp/dev_types.py
	modified:   devapp/forms.py
	modified:   devapp/locale/ru/LC_MESSAGES/django.po
	new file:   devapp/migrations/0001_squashed_0005_device_ip_address_change.py
	modified:   devapp/models.py
	new file:   devapp/onu_config/__init__.py
	new file:   devapp/onu_config/base.py
	new file:   devapp/onu_config/f601.py
	new file:   devapp/onu_config/f660.py
	modified:   devapp/templates/devapp/custom_dev_page/generic_switch.html
	modified:   devapp/templates/devapp/custom_dev_page/onu.html
	modified:   devapp/templates/devapp/custom_dev_page/onu_for_zte.html
	modified:   devapp/templates/devapp/group_list.html
	modified:   devapp/urls.py
	modified:   devapp/views.py
	modified:   dialing_app/templates/index.html
	modified:   dialing_app/templatetags/telephone_filters.py
	modified:   djing/fields.py
	modified:   djing/lib/decorators.py
	modified:   djing/lib/mixins.py
	deleted:    djing/lib/tln/__init__.py
	deleted:    djing/lib/tln/tln.py
	modified:   djing/local_settings.py.example
	modified:   djing/settings.py
	modified:   djing/tasks.py
	modified:   djing/urls.py
	modified:   djing/views.py
	modified:   docs/extra_func.md
	modified:   group_app/views.py
	modified:   gw_app/nas_managers/core.py
	modified:   gw_app/nas_managers/mod_mikrotik.py
	modified:   gw_app/nas_managers/structs.py
	new file:   ip_pool/migrations/0001_squashed_0004_auto_20190305_1243.py
	new file:   ip_pool/migrations/0004_auto_20190305_1243.py
	modified:   ip_pool/models.py
	modified:   locale/ru/LC_MESSAGES/django.po
	new file:   messenger/__init__.py
	new file:   messenger/admin.py
	new file:   messenger/apps.py
	new file:   messenger/forms.py
	new file:   messenger/locale/ru/LC_MESSAGES/django.po
	new file:   messenger/migrations/0001_initial.py
	renamed:    chatbot/__init__.py -> messenger/migrations/__init__.py
	new file:   messenger/models.py
	new file:   messenger/tasks.py
	new file:   messenger/templates/messenger/add_messenger.html
	new file:   messenger/templates/messenger/messenger_list.html
	new file:   messenger/templates/messenger/vibermessenger_form.html
	new file:   messenger/tests.py
	new file:   messenger/urls.py
	new file:   messenger/views.py
	modified:   msg_app/forms.py
	modified:   msg_app/models.py
	modified:   msg_app/templates/msg_app/chat.html
	modified:   msg_app/urls.py
	modified:   msg_app/views.py
	modified:   periodic.py
	modified:   requirements.txt
	modified:   searchapp/templates/searchapp/index.html
	modified:   static/bad_ie.html
	modified:   static/css/all.min.css
	deleted:    static/css/bootstrap-datetimepicker.min.css
	new file:   static/css/chartist.min.css
	deleted:    static/img/gmarkers/dev.png
	deleted:    static/img/gmarkers/dev_bug.png
	deleted:    static/img/gmarkers/dev_ok.png
	deleted:    static/img/gmarkers/flag_black.png
	deleted:    static/img/gmarkers/relay_rack.png
	modified:   static/js/all.min.js
	deleted:    static/js/bootstrap-datetimepicker.min.js
	new file:   static/js/chartist.min.js
	deleted:    static/js/cidr.js
	modified:   static/js/my.js
	deleted:    systemd_units/djing_telebot.service
	modified:   systemd_units/do_backup.sh
	modified:   taskapp/forms.py
	modified:   taskapp/handle.py
	modified:   taskapp/locale/ru/LC_MESSAGES/django.po
	modified:   taskapp/models.py
	modified:   taskapp/templates/taskapp/details.html
	modified:   taskapp/templates/taskapp/footer_btns.html
	modified:   taskapp/urls.py
	modified:   taskapp/views.py
	deleted:    telebot.py
	modified:   templates/all_base.html
	modified:   templates/base.html
	renamed:    chatbot/migrations/__init__.py -> traf_stat/__init__.py
	new file:   traf_stat/admin.py
	new file:   traf_stat/apps.py
	new file:   traf_stat/fields.py
	new file:   traf_stat/migrations/0001_initial.py
	new file:   traf_stat/migrations/__init__.py
	new file:   traf_stat/models.py
	new file:   traf_stat/templates/statistics/index.html
	new file:   traf_stat/tests.py
	new file:   traf_stat/urls.py
	new file:   traf_stat/views.py

+++ abonapp/migrations/0002_auto_20180808_1448.py
@@ -19,12 +19,15 @@ TMP_FILE = '/tmp/djing_ip_field_abonapp_migrate.json'
 def backup_info(apps, _):
     Abon = apps.get_model('abonapp', 'Abon')
     obs = Abon.objects.exclude(ip_address=None).only('ip_address', 'is_dynamic_ip')
-    with open(TMP_FILE, 'w') as f:
-        serializers.serialize('json', obs, stream=f, fields=('ip_address', 'is_dynamic_ip'))
+    if obs.exists():
+        with open(TMP_FILE, 'w') as f:
+            serializers.serialize('json', obs, stream=f, fields=('ip_address', 'is_dynamic_ip'))
 
 
 def restore_info_to_new_scheme(apps, _):
     Abon = apps.get_model('abonapp', 'Abon')
+    if not os.path.isfile(TMP_FILE):
+        return
     with open(TMP_FILE, 'r') as f:
         for abon in load(f):
             ip_addr = abon['fields'].get('ip_address')
